@@ -3,64 +3,62 @@
   "author": "Joe McIntyre (j_h_mcintyre@yahoo.com)",
   "description": "Stock Tracker",
   "version": "1.0.0",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/fcc-joemcintyre/stocktracker"
   },
   "license": "MIT",
   "dependencies": {
-    "body-parser": "1.18.2",
-    "d3": "4.13.0",
-    "express": "4.16.2",
-    "prop-types": "15.6.0",
-    "react": "16.2.0",
-    "react-dom": "16.2.0",
-    "request": "2.83.0",
-    "socket.io": "2.0.4",
-    "socket.io-client": "2.0.4"
+    "d3": "5.16.0",
+    "express": "4.17.1",
+    "helmet": "3.23.3",
+    "node-fetch": "2.6.0",
+    "prop-types": "15.7.2",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-is": "16.13.1",
+    "socket.io": "2.3.0",
+    "socket.io-client": "2.3.0"
   },
   "scripts": {
     "test": "jest",
-    "integration-test": "mocha test/server/test-main.js",
+    "integration-test": "mocha test/server",
     "start": "nodemon dist/main.js --port=3000",
-    "build": "gulp",
-    "build-stage": "gulp stage",
-    "lint": "eslint --ext .js --ext .jsx src/** && eslint test/**"
+    "build": "webpack --config=webpack.config.cjs --env=development -w",
+    "build-stage": "webpack --config=webpack.config.cjs --env=production",
+    "lint": "eslint --ext .js src/** && eslint test/**"
   },
   "devDependencies": {
-    "@cygns/eslint-config": "2.1.0",
-    "babel-core": "6.26.0",
-    "babel-jest": "22.4.1",
-    "babel-preset-env": "1.6.1",
-    "babel-preset-react": "6.24.1",
-    "babelify": "8.0.0",
-    "browserify": "16.1.0",
-    "enzyme": "3.3.0",
-    "enzyme-adapter-react-16": "1.1.1",
-    "eslint": "4.18.1",
-    "eslint-plugin-import": "2.9.0",
-    "eslint-plugin-react": "7.7.0",
-    "gulp": "3.9.1",
-    "gulp-babel-minify": "0.3.0",
-    "gulp-cssmin": "0.2.0",
-    "gulp-gzip": "1.4.2",
-    "gulp-sass": "3.1.0",
-    "gulp-sourcemaps": "2.6.4",
-    "gulp-util": "3.0.8",
-    "jest": "22.4.2",
-    "mocha": "5.0.1",
-    "mock-socket": "7.1.0",
-    "nodemon": "1.15.1",
-    "react-test-renderer": "16.2.0",
-    "vinyl-buffer": "1.0.1",
-    "vinyl-source-stream": "2.0.0",
-    "watchify": "3.10.0"
+    "@babel/core": "7.10.5",
+    "@babel/plugin-proposal-object-rest-spread": "7.10.4",
+    "@babel/preset-env": "7.10.4",
+    "@babel/preset-react": "7.10.4",
+    "@babel/register": "7.10.5",
+    "@cygns/eslint-config": "5.0.0",
+    "babel-jest": "26.1.0",
+    "babel-loader": "8.1.0",
+    "compression-webpack-plugin": "4.0.0",
+    "copy-webpack-plugin": "6.0.3",
+    "enzyme": "3.11.0",
+    "enzyme-adapter-react-16": "1.15.2",
+    "eslint": "7.5.0",
+    "eslint-plugin-import": "2.22.0",
+    "eslint-plugin-react": "7.20.3",
+    "eslint-plugin-react-hooks": "4.0.8",
+    "jest": "26.1.0",
+    "mocha": "8.0.1",
+    "mock-socket": "9.0.3",
+    "nodemon": "2.0.4",
+    "react-test-renderer": "16.13.1",
+    "webpack": "4.43.0",
+    "webpack-cli": "3.3.12"
   },
   "engines": {
-    "node": "8.9.3"
+    "node": "^12.18"
   },
   "jest": {
-    "setupTestFrameworkScriptFile": "./jestSetup.js",
+    "setupFilesAfterEnv": ["./jestSetup.js"],
     "collectCoverageFrom": [
       "src/**/*.{js,jsx}"
     ],
