@@ -3,21 +3,23 @@
   "author": "Joe McIntyre (j_h_mcintyre@yahoo.com)",
   "description": "Stock Tracker",
   "version": "1.0.0",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/fcc-joemcintyre/stocktracker"
   },
   "license": "MIT",
   "dependencies": {
-    "body-parser": "1.18.2",
-    "express": "4.16.2",
-    "request": "2.83.0",
-    "socket.io": "2.0.4"
+    "express": "4.17.1",
+    "helmet": "3.23.3",
+    "node-fetch": "2.6.0",
+    "socket.io": "2.3.0"
   },
   "scripts": {
     "start": "node dist/main.js --port=3000"
   },
   "engines": {
-    "node": "8.9.3"
+    "node": "12.18.2",
+    "yarn": "1.22.4"
   }
 }
