@@ -1,6 +1,8 @@
-
 import React from 'react';
 import { render } from 'react-dom';
 import { App } from './App';
 
-render (<App />, document.getElementById ('app'));
+render (
+  <App />,
+  document.getElementById ('app'),
+);
