@@ -1,4 +1,4 @@
-const request = require ('request');
+import fetch from 'node-fetch';
 
 // socket.io and stock data
 let io;
@@ -10,7 +10,7 @@ let stocks = [];
  * @param {Object} _io Socket.io instance
  * @return {void}
  */
-function init (_io) {
+export function setSocket (_io) {
   io = _io;
   io.on ('connection', (socket) => {
     console.log ('added connection');
@@ -24,9 +24,9 @@ function init (_io) {
  * @param {Object} res Response
  * @return {void}
  */
-function registerStock (req, res) {
+export async function registerStock (req, res) {
   let result;
-  let symbol = req.params.symbol;
+  let { symbol } = req.params;
   if (symbol) {
     // add if symbol not already active
     symbol = symbol.toString ().trim ().toUpperCase ();
@@ -37,8 +37,8 @@ function registerStock (req, res) {
         break;
       }
     }
-    if (! found) {
-      addStock (symbol);
+    if (!found) {
+      await addStock (symbol);
     }
     result = { errorCode: 0 };
   } else {
@@ -53,9 +53,9 @@ function registerStock (req, res) {
  * @param {Object} res Response
  * @return {void}
  */
-function deregisterStock (req, res) {
+export function deregisterStock (req, res) {
   let result;
-  let symbol = req.params.symbol;
+  let { symbol } = req.params;
   if (symbol) {
     // if present, remove from stocks
     symbol = symbol.toString ().trim ().toUpperCase ();
@@ -80,7 +80,13 @@ function deregisterStock (req, res) {
  * @return {void}
  */
 function addStock (symbol) {
-  process.nextTick (() => {
+  const key = process.env.QKEY;
+  if (!key) {
+    console.log ('Quandl key not set up');
+    return;
+  }
+
+  process.nextTick (async () => {
     // calculate start / end dates (3 year history)
     const date = new Date ();
     date.setFullYear (date.getFullYear () - 3);
@@ -88,48 +94,37 @@ function addStock (symbol) {
     const dates = `start_date=${startDate}`;
 
     // construct url and retrieve data
-    const key = process.env.QKEY;
     const keyParam = (key) ? `&api_key=${key}` : '';
     const base = `https://www.quandl.com/api/v3/datasets/WIKI/${symbol}.json`;
     const url = `${base}?order=asc&${dates}${keyParam}`;
-    request.get (url, (err, res, body) => {
-      if (! err) {
-        if (res.statusCode === 200) {
-          const data = JSON.parse (body);
-          const index = data.dataset.name.indexOf ('(');
-          const name = (index === -1) ? data.dataset.name : data.dataset.name.substring (0, index - 1);
-          // add stock to set of tracked stocks, broadcast to connected clients
-          stocks.push ({
-            status: 0,
-            symbol,
-            name,
-            data: data.dataset.data,
-          });
-          broadcast ();
-          console.log (`${symbol} now being tracked`);
-        } else {
-          stocks.push ({
-            status: res.statusCode,
-            symbol,
-            name: null,
-            data: null,
-          });
-          console.log ('Fetch error: status code ', res.statusCode);
-          const data = JSON.parse (body);
-          console.log ('  Detailed msg', data);
-          broadcast ();
-        }
-      } else {
-        stocks.push ({
-          status: 1,
-          symbol,
-          name: null,
-          data: null,
-        });
-        console.log ('Fetch error:', err);
-        broadcast ();
-      }
+    const res = await fetch (url, {
+      method: 'GET',
     });
+    const data = await res.json ();
+    if (res.ok) {
+      const index = data.dataset.name.indexOf ('(');
+      const name = (index === -1) ? data.dataset.name : data.dataset.name.substring (0, index - 1);
+      // add stock to set of tracked stocks, broadcast to connected clients
+      stocks.push ({
+        status: 0,
+        symbol,
+        name,
+        data: data.dataset.data,
+      });
+      broadcast ();
+      console.log (`${symbol} now being tracked`);
+      // fs.writeFileSync (`${symbol}.json`, JSON.stringify (data.dataset.data));
+    } else {
+      stocks.push ({
+        status: res.statusCode,
+        symbol,
+        name: null,
+        data: null,
+      });
+      console.log ('Fetch error: status code ', res.statusCode);
+      console.log ('  Detailed msg', data);
+      broadcast ();
+    }
   });
 }
 
@@ -144,11 +139,7 @@ function broadcast () {
       io.emit ('update', JSON.stringify (stocks));
 
       // clean up any error records after each broadcast
-      stocks = stocks.filter (a => (a.status === 0));
+      stocks = stocks.filter ((a) => a.status === 0);
     }
   });
 }
-
-exports.init = init;
-exports.registerStock = registerStock;
-exports.deregisterStock = deregisterStock;
