@@ -1,12 +1,11 @@
+import * as listener from './listener.js';
+
 /**
  * Initialize routes.
  * @param {Object} app Express instance
- * @param {Object} listener Route implementations
  * @return {void}
  */
-function init (app, listener) {
+export function init (app) {
   app.put ('/api/stocks/:symbol', listener.registerStock);
   app.delete ('/api/stocks/:symbol', listener.deregisterStock);
 }
-
-exports.init = init;
