@@ -1,5 +1,5 @@
-const request = require ('request');
-const server = require ('../../dist/server');
+import fetch from 'node-fetch';
+import * as server from '../../src/server/server.js';
 
 before (function () {
   server.start (3999);
@@ -11,64 +11,70 @@ after (async function () {
 
 describe ('test server', function () {
   describe ('/', function () {
-    it ('should return 200 with home page', function (done) {
-      request.get ('http://localhost:3999/', function (err, res, body) {
-        if (err) { return done (err); }
-        if (res.statusCode === 200) {
-          if (body.indexOf ('<title>Stock Tracker</title>') !== -1) {
-            return done ();
-          } else {
-            return (done (new Error ('Invalid body', body)));
-          }
+    it ('should return 200 with home page', async () => {
+      const res = await fetch ('http://localhost:3999/', { method: 'GET' });
+      if (!res.ok) {
+        throw new Error ('Invalid request');
+      }
+      if (res.status === 200) {
+        const data = await res.text ();
+        if (data.indexOf ('<title>Stock Tracker</title>') === -1) {
+          throw new Error ('Invalid body', data);
         }
-        return done (new Error ('Invalid response', res.statusCode));
-      });
+      } else {
+        throw new Error ('Invalid response', res.status);
+      }
     });
   });
 
   describe ('invalid URL content', function () {
-    it ('should return 200 with instructions', function (done) {
-      request.get ('http://localhost:3999/dummy', function (err, res) {
-        if (err) { return done (err); }
-        if (res.statusCode === 404) {
-          return done ();
+    it ('should return 200 with home page', async () => {
+      const res = await fetch ('http://localhost:3999/dummy', { method: 'GET' });
+      if (!res.ok) {
+        throw new Error ('Invalid request');
+      }
+      if (res.status === 200) {
+        const data = await res.text ();
+        if (data.indexOf ('<title>Stock Tracker</title>') === -1) {
+          throw new Error ('Invalid body', data);
         }
-        return done (new Error ('Invalid response', res.statusCode));
-      });
+      } else {
+        throw new Error ('Invalid response', res.status);
+      }
     });
   });
 
   describe ('valid register stock request', function () {
-    it ('should register stock', function (done) {
-      request.put ('http://localhost:3999/api/stocks/AAPL', function (err, res, body) {
-        if (err) { return done (err); }
-        if (res.statusCode === 200) {
-          const data = JSON.parse (body);
-          if (data.errorCode === 0) {
-            return done ();
-          } else {
-            return done (new Error (`Invalid errorCode ${data.errorCode}`));
-          }
+    it ('should register stock', async () => {
+      const res = await fetch ('http://localhost:3999/api/stocks/AAPL', { method: 'PUT' });
+      if (!res.ok) {
+        throw new Error ('Invalid request');
+      }
+      if (res.status === 200) {
+        const data = await res.json ();
+        if (data.errorCode !== 0) {
+          throw new Error ('Invalid errorCode', data.errorCode);
         }
-        return done (new Error (`Invalid status code ${res.statusCode}`));
-      });
+      } else {
+        throw new Error ('Invalid response', res.status);
+      }
     });
   });
 
   describe ('valid deregister stock request', function () {
-    it ('should deregister stock', function (done) {
-      request.del ('http://localhost:3999/api/stocks/AAPL', function (err, res, body) {
-        if (err) { return done (err); }
-        if (res.statusCode === 200) {
-          const data = JSON.parse (body);
-          if (data.errorCode === 0) {
-            return done ();
-          } else {
-            return done (new Error (`Invalid errorCode ${data.errorCode}`));
-          }
+    it ('should deregister stock', async () => {
+      const res = await fetch ('http://localhost:3999/api/stocks/AAPL', { method: 'DELETE' });
+      if (!res.ok) {
+        throw new Error ('Invalid request');
+      }
+      if (res.status === 200) {
+        const data = await res.json ();
+        if (data.errorCode !== 0) {
+          throw new Error ('Invalid errorCode', data.errorCode);
         }
-        return done (new Error (`Invalid status code ${res.statusCode}`));
-      });
+      } else {
+        throw new Error ('Invalid response', res.status);
+      }
     });
   });
 });
